发布1.0.3,修复错误并增加部分功能

+++ opencode-base-graph/src/main/java/cloud/opencode/base/graph/algorithm/BidirectionalBfsUtil.java
@@ -271,7 +271,7 @@ public final class BidirectionalBfsUtil {
 
         // Filter by total distance
         result.removeIf(v -> {
-            int totalDist = distFromSource.getOrDefault(v, Integer.MAX_VALUE)
+            long totalDist = (long) distFromSource.getOrDefault(v, Integer.MAX_VALUE)
                 + distFromTarget.getOrDefault(v, Integer.MAX_VALUE);
             return totalDist > maxDistance;
         });
